beginning resolving; BROKEN
