+++ package.json
@@ -50,6 +50,7 @@
     "storybook": "^7.6.17",
     "style-loader": "^3.3.3",
     "tailwindcss": "^3.4.1",
+    "ts-interface-builder": "^0.3.3",
     "ts-loader": "^9.4.4",
     "typescript": "^4.9.5"
   },
@@ -75,6 +76,7 @@
     "match-sorter": "^6.3.1",
     "next": "^14.1.0",
     "react-csv": "^2.2.2",
-    "react-router-dom": "^6.19.0"
+    "react-router-dom": "^6.19.0",
+    "ts-interface-checker": "^1.0.2"
   }
 }
