+++ src/components/Table/__deprecated__/sorting.tsx
@@ -5,7 +5,7 @@
 
 import { Row, SortingFn } from "@tanstack/react-table";
 import { parseFieldValue } from "../utils";
-import { TableData } from "../deprecated/tables";
+import { TableData } from "./tables";
 
 const reSplitAlphaNumeric = /([0-9]+)/gm;
